Fix gcc-only warnings and intrusive_ptr type punning
gcc 11 and 13 report what Apple clang does not, and CI builds under
-Werror:
- -Wcomment: a commented-out macro in nodes.cpp ended in a backslash;
- -Wsign-promo: uint8_t/uint16_t passed to std::to_string and to
  absl::int128 (utils.cpp, transform_insert.cpp);
- -Wstrict-aliasing: A_Const::val read through nodeTag(&val)
  (transform_select.cpp), and intrusive_ptr<base> read as
  intrusive_ptr<derived> through reinterpret_cast<derived_ptr&>.

Those reinterpret_cast reads were undefined behaviour at every site, not
only where gcc 13 flagged them, so all 45 are rewritten:
static_cast<T*>(p.get()) where only -> is used, and
boost::static_pointer_cast where the pointer is passed on.

+++ components/sql/transformer/impl/transform_insert.cpp
@@ -138,9 +138,9 @@ namespace {
                 case LT::BIGINT:
                     return components::types::logical_value_t(resource, int128_t{val.value<int64_t>()});
                 case LT::UTINYINT:
-                    return components::types::logical_value_t(resource, int128_t{val.value<uint8_t>()});
+                    return components::types::logical_value_t(resource, int128_t{static_cast<uint32_t>(val.value<uint8_t>())});
                 case LT::USMALLINT:
-                    return components::types::logical_value_t(resource, int128_t{val.value<uint16_t>()});
+                    return components::types::logical_value_t(resource, int128_t{static_cast<uint32_t>(val.value<uint16_t>())});
                 case LT::UINTEGER:
                     return components::types::logical_value_t(resource, int128_t{val.value<uint32_t>()});
                 case LT::UBIGINT:
